+++ src/Controller/RoadmapGroupController.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Controller\Trait\ModuleGatedControllerTrait;
 use App\Entity\RoadmapGroup;
 use App\Form\RoadmapGroupType;
 use App\Repository\RoadmapGroupRepository;
+use App\Service\ModuleConfigurationService;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Bundle\SecurityBundle\Security;
@@ -18,17 +20,22 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 
 class RoadmapGroupController extends AbstractController
 {
+    use ModuleGatedControllerTrait;
+
     public function __construct(
         private readonly RoadmapGroupRepository $roadmapGroupRepository,
         private readonly EntityManagerInterface $entityManager,
         private readonly TranslatorInterface $translator,
-        private readonly Security $security
+        private readonly Security $security,
+        private readonly ModuleConfigurationService $moduleService,
     ) {}
 
     #[Route('/planning/groups', name: 'app_planning_group_index', methods: ['GET'])]
     #[IsGranted('ROLE_USER')]
     public function index(): Response
     {
+        if ($redirect = $this->checkModuleActive('resource_planning')) return $redirect;
+
         $tenant = $this->security->getUser()?->getTenant();
         $groups = $tenant ? $this->roadmapGroupRepository->findActiveByTenant($tenant) : [];
 
@@ -41,6 +48,8 @@ class RoadmapGroupController extends AbstractController
     #[IsGranted('ROLE_MANAGER')]
     public function new(Request $request): Response
     {
+        if ($redirect = $this->checkModuleActive('resource_planning')) return $redirect;
+
         $group = new RoadmapGroup();
         $tenant = $this->security->getUser()?->getTenant();
         if ($tenant !== null) {
@@ -72,6 +81,8 @@ class RoadmapGroupController extends AbstractController
     #[IsGranted('ROLE_MANAGER')]
     public function edit(Request $request, RoadmapGroup $roadmapGroup): Response
     {
+        if ($redirect = $this->checkModuleActive('resource_planning')) return $redirect;
+
         $form = $this->createForm(RoadmapGroupType::class, $roadmapGroup);
         $form->handleRequest($request);
 
@@ -96,6 +107,8 @@ class RoadmapGroupController extends AbstractController
     #[IsGranted('ROLE_MANAGER')]
     public function delete(Request $request, RoadmapGroup $roadmapGroup): Response
     {
+        if ($redirect = $this->checkModuleActive('resource_planning')) return $redirect;
+
         if ($this->isCsrfTokenValid('delete'.$roadmapGroup->getId(), $request->request->get('_token'))) {
             $this->entityManager->remove($roadmapGroup);
             $this->entityManager->flush();
